Create main-model
The main model for our analysis uses the #n time series. These time series are listed in the following table with their relationship to raw CRUDE counterparts. The variable Y contains the main time series for the model and the variable iY contains integrated data from Y that will be used in forecasting analyses. The date range for available data is xx-19xx to Feb-2016.